remove _dntEnabled from global scope
update dnt-helper.js from source, move into libs folder
